+++ backend/bookmarks/urls.py
@@ -1,8 +1,15 @@
 from django.urls import path
 
-from .views import BookmarksListView, bookmark_create_from_telegram
+from .views import (
+    TagListView,
+    FolderListView,
+    BookmarkListView,
+    bookmark_create_from_telegram,
+)
 
 urlpatterns = [
-    path("bookmarks/", BookmarksListView.as_view()),
+    path("tags/", TagListView.as_view()),
+    path("folders/", FolderListView.as_view()),
+    path("bookmarks/", BookmarkListView.as_view()),
     path("bookmarks/create_from_telegram/", bookmark_create_from_telegram),
 ]